refactor(analysis-types): extract baseline aggregate DTO
